+++ lib/app_publisher/android_publisher_command.dart
@@ -1,19 +1,18 @@
-import 'package:distribute_cli/app_publisher/fastlane/fastlane_android_command.dart';
-import 'package:distribute_cli/app_publisher/firebase/firebase_android_command.dart';
+import 'package:distribute_cli/app_publisher/fastlane/android/command.dart' as fastlane_publisher;
+import 'package:distribute_cli/app_publisher/firebase/android/command.dart' as firebase_publisher;
 
 import '../command.dart';
+import 'github/command.dart' as github;
 
 /// A command to publish an Android application.
 ///
 /// The `AndroidPublisherCommand` class provides subcommands for publishing
 /// Android apps using different tools, such as Fastlane and Firebase.
 class AndroidPublisherCommand extends Commander {
-  /// Creates a new `AndroidPublisherCommand` instance.
-  ///
-  /// Adds subcommands for Fastlane and Firebase publishers.
   AndroidPublisherCommand() {
-    addSubcommand(FastlaneAndroidCommand());
-    addSubcommand(FirebaseAndroidCommand());
+    addSubcommand(fastlane_publisher.Command());
+    addSubcommand(firebase_publisher.Command());
+    addSubcommand(github.Command("android"));
   }
 
   /// A description of the command.
@@ -28,5 +27,7 @@ class AndroidPublisherCommand extends Commander {
   ///
   /// This method is currently not implemented.
   @override
-  Future? run() async {}
+  Future? run() async {
+    super.run();
+  }
 }