+++ lib/parsers/job_arguments.dart
@@ -1,3 +1,9 @@
+import 'package:distribute_cli/app_builder/android/arguments.dart' as android_arguments;
+
+import '../app_builder/ios/arguments.dart' as ios_arguments;
+import '../app_publisher/fastlane/android/arguments.dart' as fastlane_publisher;
+import '../app_publisher/firebase/android/arguments.dart' as firebase_publisher;
+import '../app_publisher/xcrun/ios/arguments.dart' as xcrun_publisher;
 import 'task_arguments.dart';
 
 enum JobMode {
@@ -18,12 +24,64 @@ enum JobMode {
 
 abstract class JobArguments {
   List<String> get results;
-  List<String> get argKeys;
-  JobMode get jobMode;
-
-  late Job parent;
 
   Map<String, dynamic> toJson();
+}
+
+class BuilderJob {
+  final android_arguments.Arguments? android;
+  final ios_arguments.Arguments? ios;
+  late Job parent;
+
+  BuilderJob({this.android, this.ios}) {
+    if (android == null && ios == null) {
+      throw Exception("Android or iOS build argument must be provided.");
+    }
+    android?.parent = this;
+    ios?.parent = this;
+  }
+
+  factory BuilderJob.fromJson(Map<String, dynamic> json) {
+    return BuilderJob(
+      android: json["android"] != null ? android_arguments.Arguments.fromJson(json["android"]) : null,
+      ios: json["ios"] != null ? ios_arguments.Arguments.fromJson(json["ios"]) : null,
+    );
+  }
+
+  Map<String, dynamic> toJson() => {
+        if (android != null) "android": android?.toJson(),
+        if (ios != null) "ios": ios?.toJson(),
+      };
+}
+
+class PublisherJob {
+  final fastlane_publisher.Arguments? fastlane;
+  final firebase_publisher.Arguments? firebase;
+  final xcrun_publisher.Arguments? xcrun;
+  late Job parent;
+
+  PublisherJob({this.fastlane, this.firebase, this.xcrun}) {
+    if (fastlane == null && xcrun == null && firebase == null) {
+      throw Exception("Fastlane, Firebase, or iOS publisher argument must be provided.");
+    }
+    fastlane?.parent = this;
+    firebase?.parent = this;
+    xcrun?.parent = this;
+  }
+
+  Map<String, dynamic> toJson() => {
+        if (fastlane != null) "fastlane": fastlane?.toJson(),
+        if (firebase != null) "firebase": firebase?.toJson(),
+        if (xcrun != null) "xcrun": xcrun?.toJson(),
+      };
+
+  factory PublisherJob.fromJson(Map<String, dynamic> json) {
+    return PublisherJob(
+      fastlane: json["fastlane"] != null ? fastlane_publisher.Arguments.fromJson(json["fastlane"]) : null,
+      firebase: json["firebase"] != null ? firebase_publisher.Arguments.fromJson(json["firebase"]) : null,
+      xcrun: json["xcrun"] != null ? xcrun_publisher.Arguments.fromJson(json["xcrun"]) : null,
+    );
+  }
 }
 
 /// Represents a job in the configuration.
@@ -31,29 +89,23 @@ abstract class JobArguments {
 /// A `Job` consists of a name, an optional key, an optional description,
 /// a platform, a mode (build or publish), a package name, and associated arguments.
 class Job {
-  /// The name of the job.
-  final String name;
-
   /// The unique key of the job (optional).
   final String? key;
 
+  /// The name of the job.
+  final String name;
+
+  final PublisherJob? publisher;
+  final BuilderJob? builder;
+
   /// The description of the job (optional).
   final String? description;
-
-  /// The platform for which the job is executed (e.g., android, ios).
-  final String platform;
-
-  /// The mode of the job (build or publish).
-  final JobMode mode;
 
   /// The package name associated with the job.
   final String packageName;
 
   /// The environment variables for the job (optional).
   final Map<String, dynamic>? environments;
-
-  /// The arguments associated with the job.
-  final JobArguments arguments;
 
   /// The parent task of the job.
   late Task parent;
@@ -72,47 +124,36 @@ class Job {
     required this.name,
     this.key,
     required this.description,
-    required this.arguments,
     required this.packageName,
-    required this.platform,
-    required this.mode,
     this.environments,
-  }) {
-    arguments.parent = this;
+    this.builder,
+    this.publisher,
+  }) : assert(
+          (builder != null && publisher == null) || (builder == null && publisher != null),
+          "Either builder or publisher must be provided, not both.",
+        ) {
+    if (builder != null) {
+      builder?.parent = this;
+    } else if (publisher != null) {
+      publisher?.parent = this;
+    } else {
+      throw Exception("Either builder or publisher must be provided.");
+    }
   }
 
   factory Job.fromJson(Map<String, dynamic> json) {
-    final platform = json["platform"];
-    final mode = json["mode"];
     final packageName = json["package_name"];
     final key = json["key"];
-
-    if (mode == null) throw Exception("mode is required for each job");
-    if (mode != "build" && mode != "publish") {
-      throw Exception("Invalid mode for each job");
-    }
-    if (platform == null) throw Exception("platform is required for each job");
     if (packageName == null) {
       throw Exception("package_name is required for each job");
-    }
-
-    JobArguments? task;
-    final isBuildMode = mode == "build";
-    final isPublishMode = mode == "publish";
-
-    if (!isBuildMode && !isPublishMode) {
-      throw Exception("Invalid job mode");
     }
 
     return Job(
       name: json["name"],
       key: key,
       description: json["description"],
-      arguments: task!,
       environments: json["variables"],
       packageName: packageName,
-      platform: platform,
-      mode: JobMode.fromString(mode),
     );
   }
 
@@ -122,8 +163,7 @@ class Job {
         "key": key,
         "description": description,
         "package_name": packageName,
-        "platform": platform,
-        "mode": mode.name,
-        "arguments": arguments.toJson(),
+        if (builder != null) "builder": builder?.toJson(),
+        if (publisher != null) "publisher": publisher?.toJson(),
       };
 }